refactor: lift validateRepoName into shared utils/git module
